Conflicts:
	salvus/admin.py

+++ salvus/vm.py
@@ -50,6 +50,9 @@ def run_kvm(ip_address, hostname, vcpus, ram):
 
                 #### hostname ####
                 hostname_file = os.path.join(tmp_path,'etc/hostname')
+                if not os.path.exists(hostname_file):
+                    raise RuntimeError("missing /etc/hostname in the VM image-- the guestmount command is not working; run with verbose directly, check for permission issues, etc.")
+                    
                 os.unlink(hostname_file)
                 open(hostname_file,'w').write(hostname)
                 hosts_file = os.path.join(tmp_path, 'etc/hosts')
